Allow more flexible arguments for `draw_tree`
This lets the outer parens be dropped when passing in a list-constructed
tree structure, and as a side-effect, prevents someone who would have
made this mistake before from accidentally providing a string for the
options parameter (now options always needs to be a named parameter). It
should behave the same for usual nltk situations.