Handle very long non-leaf nodes
Previously, internal node width was ignored, which leads to bad
rendering when the leaf nodes are very short and internal labels are
very long. This now takes the max of the subtree widths and the label
width, with a minimum width of 1em.

Also, adjust some variable names to be less confusing.