Support multi-line leaf nodes (pace Andrew Carnie)
This doesn't handle multi-line nodes in a very general way, but the
change for leaf nodes is easy enough to make. This commit also
introduces the ability to pass option parameters simply as named
arguments to `draw_tree`, which is a lot less verbose.

Also, documentation improvements and a few style fixes.