New, more general, layout engine
This commit provides a new renderer that is designed from the beginning
to handle multi-line nodes, as well as to retain layout info after the
initial rendering. The renderer is still in testing, but can do
everything in the documentation notebook plus multi-line nodes in
tree-internal positions. It is also missing a few options, to be added
later. The stored layout can be used for adding later annotations to
trees, in particular, this is a precondition for movement arrows (which
are not yet implemented).