Add some basic annotations on top of trees
This code allows adding extra annotations on top of the regular tree,
along with a bunch of support code that let you calculate node
positioning relative to the outermost svg canvas using tree paths. The
first two kinds of annotations introduced here are pretty simple:
drawing a box around a constituent, and drawing a line under a
constituent.