Make svgling diagrams more responsive
This implements viewbox and preserveAspectRatio on all main figure
elements, allowing for cleaner resizing. All internal units are changed
to user units, except for the width/height for main diagrams (trees and
complex figures), which report the same values in px. The end result is
that, as before, 1 user unit defaults to 1px, but resizing actually
works. One small note is that, as far as I understand it, css font-size
on svg is required to have a unit, so px for this value is interpreted
as user units rather than true px.

The most immediate impact of this change is that nbconvert outputs
(which force scaling of over-wide diagrams) will not be clipped when a
diagram doesn't fit, e.g. on a mobile device. For the most part,
interactive rendering should be unchanged (since interactive notebooks
let svg diagrams scroll).