Initial support for hybrid HTML/CSS/SVG trees
These diagrams use divs + css for layout (mostly using css grids), with
inline SVG for drawing the lines in the trees. They require fairly
recent browsers to work at all.

These diagrams are fairly limited compared to pure SVG diagrams, but
they provide a robust way of doing certain kinds of trees that involve
arbitrary HTML nodes. There are a few limitations that ought to be
fixed, but for all of them the fix is not obvious: (i) branching can be
no more than binary (due to the trick used to mix HTML and SVG), (ii)
the spacing of parent nodes isn't calculated right (due to the trick
used to get them centered relative to the tree lines), and (iii) these
still aren't rendering as intended in current Safari. I'm not sure why
(iii) is happening except that it seems to be that Safari makes
different choices about how to scale SVGs that do not preserve aspect
ratios than Firefox/Chrome. The main impact is that for smaller trees
(relative to the window) a lot of whitespace is added, so it's not
crippling.