Provide a shortcut for multiline nodes, documentation tweaks
Because MathJax multiline support seems to be unreliable across
rendering engines, provide a quick way to do multi-line text that
doesn't rely on this and convert the documentation / gallery examples so
that they don't rely on MathJax for this part. Also update and add to
the documentation in various ways.