feat: add a mode that does not use `em`s
This mode is aimed at compatibility with Inkscape, which doesn't seem to
handle the combination of nested svgs and `em`s very well (as far as I
can tell it is supposed to support `em`s). See discussion in #4. I'm not
very confident that this will work right in all cases so it's not on by
default, but it does work fine for all my test and demo examples.

This commit also updates the docs to provide a bit more information
about compatibility. Along the way I fixed some doc issues to do with
current versions of MathJax, and reran the documentation notebooks.