fix: remove outdated nltk monkeypatching
As of quite a while ago now, nltk supports using svgling directly and
has also removed its `_repr_png_()`, so none of this is needed except
perhaps for backwards compatibility reasons. I've kept the functions
around for that reason, but we now rely on the built in
`nltk.Tree._repr_svg_()` to handle tree rendering.

Also, adjust documentation accordingly. (A few miscellaneous typos and
edits during the process.)