refactor: generalize TreeOptions a bit
This converts this class to a mapping with a fixed set of keys; the keys
are still stored on the class object as well. This is intended to reduce
code duplication and make it easier to go to a dict and back; as well as
to copy/initialize from TreeOptions in a non-ad-hoc way.