feat: implement basic per-node style in svgling.core
This is a longstanding feature request, but because my layout algorithm
is pretty weird and not designed for this, it wasn't all that easy to do
- handling different font sizes in particular is painful. (I'm not sure
in the end all this precalculation is worth it, but it'd take a major
rewrite to do anything else.) As of right now, only font style and size
(and debug mode) are allowed to be set, but I'm guessing these are the
main use cases (manual spacing adjustments might be nice too).

Resolves #5 (though I do realize that issue mentioned colors, which I
will also add sooner or later -- when I do add them, it'll be fairly
automatic to do them per-node).