feat: wrappers for conversion to png/pdf/ps using cairosvg
A question I see about svgling is how to convert to a raster format. My
answer in the past has essentially been "use cairosvg" (see #4). This
commit adds some light wrapper functions to make this easier, and
documents it in the manual. I've also for now  marked cairosvg as an
optional dependency (it's pretty heavy as a full dep, and maybe people
don't always care about this).