fix: remove an undeclared dependency on IPython
Noticed in env testing on release. I think I had probably added this
import in the distant past because `draw_tree` itself is designed to be
called in an IPython context and the results are a bit uninterpretable
if it isn't --  but it doesn't actually require IPython to work, and the
package setup doesn't even have ipython as an optional dep. To make this
situation clearer I also added a wrapper on top of `draw_tree` that gets
the repr directly, in case someone does want to use this easier api
programmatically without ipython.

I think most people do use this package with IPython, so this hopefully
isn't a big deal. TreeLayout direct instantiation would also still work,
which is what `__main__.py` used.