docs: first pass conversion to a quarto website
This converts the package documentation into a quarto project that
generates a website. The individual notebooks (now renamed) are still
fine to run standalone without quarto, but the whole thing now goes
together as a coherent website.

Lots of incidental document edits in the process. At the moment, there
aren'te good links between docs aside from the navbar.  This commit also
does various changelog tweaking to get it to render better as a part of
the website, including inverting the order so that it is reverse
chronological.