feat: support user-supplied `tree_split` functions
This provides an easy way to write custom tree object handling code; the
immediate market for this is nltk's ProbabilisticTree class (which does
not store the probability in the label), and a custom split function
`svgling.core.probtree_split` is provided but not used by default.  It's
more generally useful for doing complex rendering, especially for html
trees.