docs: document `tree_split` option
This updates the manual to document the new `tree_split` option, showing
how to use it for nltk's ProbabilisticTree.

This commit includes a more complex example of this in the gallery, with
actual parsing from a PCFG.  However, I've had some fairly annoying
technical issues getting this example to render in quarto, and so for
now it is not shown in the rendered gallery.