feat: further enhance custom node interface
This commit provides much more robust handling for custom for custom
node building begun in 087404963a, in particular providing a way to
defer application of tree options until the tree is built, even if the
node is instantiated as part of a tree description. It also provides a
proof of concept node builder the implements subscripts, which is
probably the most needed special formatting tweak for syntax trees.

In some ways it's a bit tempting to go full on latex math mode parsing
here, but unfortunately there don't seem to be any off-the-shelf
solutions in python that aren't extremely heavy (e.g. calling latex
itself, calling katex via node.js, etc). And it's probably not a good
use of my time to reimplement mathjax/katex...