fix: restore the use of descender info for margins, not padding
d69467524b6 fix a bunch of issues with how descender info is used for
calculating node height, but it the proceeded to use this info for node
padding, not for the margin. The previous approach used it for the
margin (adjusting edge starts acoordingly), which kept nodes much more
nicely aligned with an em grid. Restore the previous behavior with some
more refactoring. This also provides a better way of handling subscripts
-- they now also expand the margin a bit, rather than adding to the node
height itself.